+++ R/utils.R
@@ -1688,6 +1688,8 @@ cosine_distance = function(sentence1, sentence2, split_separator = " ") {
 #'
 #' the \emph{triplet_data} function returns the triplet data, which is used internally (in c++), to construct the Term Matrix. The triplet data could be usefull for secondary purposes, such as in word vector representations.
 #'
+#' the \emph{global_term_weights} function returns a list of length two. The first sublist includes the \emph{terms} and the second sublist the \emph{global-term-weights}. The \emph{tf_idf} parameter should be set to FALSE and the \emph{normalize} parameter to NULL. This function is normally used in conjuction with word-vector-embeddings.
+#'
 #' the \emph{Term_Matrix_Adjust} function removes sparse terms from a sparse matrix using a sparsity threshold
 #'
 #' the \emph{term_associations} function finds the associations between the given terms (Terms argument) and all the other terms in the corpus by calculating their correlation. There is also the option to keep a specific number of terms from the output table using the \emph{keep_terms} parameter.
@@ -1712,6 +1714,10 @@ cosine_distance = function(sentence1, sentence2, split_separator = " ") {
 #'  \item{\code{--------------}}{}
 #'
 #'  \item{\code{triplet_data()}}{}
+#'
+#'   \item{\code{--------------}}{}
+#'
+#'  \item{\code{global_term_weights()}}{}
 #'
 #'  \item{\code{--------------}}{}
 #'
@@ -1755,6 +1761,14 @@ cosine_distance = function(sentence1, sentence2, split_separator = " ") {
 #' #---------------
 #'
 #' # sm$triplet_data()
+#'
+#'
+#' #----------------------
+#' # global-term-weights :
+#' #----------------------
+#'
+#' # sm$global_term_weights()
+#'
 #'
 #' #-------------------------
 #' # removal of sparse terms:
@@ -1964,7 +1978,9 @@ sparse_term_matrix <- R6::R6Class("sparse_term_matrix",
                                         private$save_terms = as.vector(tmp_res$terms)
 
                                         if (sum(c("", " ") %in% private$save_terms) > 0) {
-
+                                          
+                                          cat("\n")
+                                          
                                           warning("empty character strings present in the column names they will be replaced with proper characters", call. = F)
                                         }
 
@@ -2028,8 +2044,45 @@ sparse_term_matrix <- R6::R6Class("sparse_term_matrix",
 
                                       return(lst)
                                     },
-
-
+                                    
+                                    
+                                    #---------------------------------
+                                    # returns the global-term-weights
+                                    #---------------------------------
+                                    
+                                    global_term_weights = function() {
+                                      
+                                      if (!self$document_term_matrix) {
+                                        
+                                        stop("the 'document_term_matrix' parameter should be set to TRUE because the global-term-weights are based on a document-term-matrix", call. = F)
+                                      }
+                                      
+                                      if (is.null(private$save_terms)) {
+                                        
+                                        stop("before calling the 'global_term_weights' method you should run the 'Term_Matrix' method", call. = F)
+                                      }
+                                      
+                                      tmp_lst = idf_global_term_weights(private$save_sparse_mat, private$save_terms)
+                                      
+                                      # remove duplicates which might be created due to 'make.names' function in the 'Term_Matrix' method
+                                      # it is possible that these duplicates might have a different idf
+                                      dupl_idx = Not_Duplicated(tmp_lst$terms)
+                                      
+                                      if (all(dupl_idx)) {
+                                        
+                                        return(tmp_lst)}
+                                      
+                                      else {
+                                        
+                                        tmp_terms = tmp_lst$terms[dupl_idx]
+                                        
+                                        tmp_gtm = tmp_lst$Idf_global_term_weights[dupl_idx]
+                                        
+                                        return(list(terms = tmp_terms, Idf_global_term_weights = tmp_gtm))
+                                      }
+                                    },
+                                    
+                                    
                                     #-----------------------------------------------------------------
                                     # subset the sparse matrix removing sparse terms using a threshold
                                     #-----------------------------------------------------------------